+++ QuickConnectPlugin.OptionsFormLauncher/InMemoryQuickConnectPluginSettings.cs
@@ -11,6 +11,7 @@ namespace QuickConnectPlugin.OptionsFormLauncher {
         public string WinScpPath { get; set; }
         public string HostAddressMapFieldName { get; set; }
         public string ConnectionMethodMapFieldName { get; set; }
+        public string AdditionalOptionsMapFieldName { get; set; }
 
         public void Load() {
             Debug.WriteLine(this.GetType().Name + ".Load()");