+++ QuickConnectPlugin/ArgumentsFormatters/PuttyArgumentsFormatter.cs
@@ -1,0 +1,51 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace QuickConnectPlugin.ArgumentsFormatters {
+
+    public class PuttyArgumentsFormatter : IArgumentsFormatter {
+
+        public String ExecutablePath { get; private set; }
+        public IPuttySessionFinder PuttySessionFinder { get; private set; }
+
+        public PuttyArgumentsFormatter(String puttyPath, IPuttySessionFinder puttySessionFinder) {
+            this.ExecutablePath = puttyPath;
+            this.PuttySessionFinder = puttySessionFinder;
+        }
+
+        public String Format(IHostPwEntry hostPwEntry) {
+            StringBuilder sb = new StringBuilder();
+            sb.AppendFormat("\"{0}\"", this.ExecutablePath);
+
+            PuttyOptions options = null;
+            bool success = PuttyOptionsParser.TryParse(hostPwEntry.AdditionalOptions, out options);
+
+            if (success && !String.IsNullOrEmpty(options.SessionName)) {
+                ICollection<String> sessionNames = this.PuttySessionFinder.Find(options.SessionName);
+
+                if (sessionNames.Count > 0) {
+                    sb.AppendFormat(" -load \"{0}\"", new List<String>(sessionNames)[0]);
+                }
+            }
+
+            if (hostPwEntry.ConnectionMethods.Contains(ConnectionMethodType.PuttySSH)) {
+                sb.Append(" -ssh");
+            }
+            else if (hostPwEntry.ConnectionMethods.Contains(ConnectionMethodType.PuttyTelnet)) {
+                sb.Append(" -telnet");
+            }
+
+            sb.AppendFormat(" {0}@{1}", hostPwEntry.GetUsername(), hostPwEntry.IPAddress);
+
+            // Specifying the password via -pw switch only works with SSH protocol.
+            // See: http://the.earth.li/~sgtatham/putty/0.65/htmldoc/Chapter3.html.
+            if (hostPwEntry.ConnectionMethods.Contains(ConnectionMethodType.PuttySSH)) {
+                // Allow passwords with white-spaces.
+                sb.AppendFormat(" -pw \"{0}\"", hostPwEntry.GetPassword());
+            }
+
+            return sb.ToString();
+        }
+    }
+}