+++ QuickConnectPlugin/FormOptions.cs
@@ -40,6 +40,7 @@ namespace QuickConnectPlugin {
             // Always add empty items.
             this.comboBoxHostAddressMapFieldName.Items.Add(String.Empty);
             this.comboBoxConnectionMethodMapFieldName.Items.Add(String.Empty);
+            this.comboBoxAdditionalOptionsMapFieldName.Items.Add(String.Empty);
 
             if (dbFields == null) {
                 this.labelWarningMessage.Visible = true;
@@ -52,15 +53,20 @@ namespace QuickConnectPlugin {
                 if (!String.IsNullOrEmpty(settings.ConnectionMethodMapFieldName)) {
                     this.comboBoxConnectionMethodMapFieldName.Items.Add(settings.ConnectionMethodMapFieldName);
                 }
+                if (!String.IsNullOrEmpty(settings.AdditionalOptionsMapFieldName)) {
+                    this.comboBoxAdditionalOptionsMapFieldName.Items.Add(settings.AdditionalOptionsMapFieldName);
+                }
 
                 this.comboBoxHostAddressMapFieldName.SelectedValue = settings.HostAddressMapFieldName;
                 this.comboBoxConnectionMethodMapFieldName.SelectedValue = settings.ConnectionMethodMapFieldName;
+                this.comboBoxAdditionalOptionsMapFieldName.SelectedValue = settings.AdditionalOptionsMapFieldName;
             }
             else {
                 this.labelWarningMessage.Visible = false;
                 foreach (var field in dbFields) {
                     this.comboBoxHostAddressMapFieldName.Items.Add(field);
                     this.comboBoxConnectionMethodMapFieldName.Items.Add(field);
+                    this.comboBoxAdditionalOptionsMapFieldName.Items.Add(field);
                 }
             }
 
@@ -78,6 +84,13 @@ namespace QuickConnectPlugin {
                 this.comboBoxConnectionMethodMapFieldName.SelectedIndex =
                     this.comboBoxConnectionMethodMapFieldName.FindStringExact(this.settings.ConnectionMethodMapFieldName);
             }
+            if (String.IsNullOrEmpty(settings.AdditionalOptionsMapFieldName)) {
+                this.comboBoxAdditionalOptionsMapFieldName.SelectedValue = String.Empty;
+            }
+            else {
+                this.comboBoxAdditionalOptionsMapFieldName.SelectedIndex =
+                    this.comboBoxAdditionalOptionsMapFieldName.FindStringExact(this.settings.AdditionalOptionsMapFieldName);
+            }
 
             // Add handlers.
             this.checkBoxEnable.CheckedChanged += new EventHandler(settingsChanged);
@@ -86,6 +99,7 @@ namespace QuickConnectPlugin {
             this.textBoxWinScpPath.TextChanged += new EventHandler(settingsChanged);
             this.comboBoxHostAddressMapFieldName.SelectedIndexChanged += new EventHandler(settingsChanged);
             this.comboBoxConnectionMethodMapFieldName.SelectedIndexChanged += new EventHandler(settingsChanged);
+            this.comboBoxAdditionalOptionsMapFieldName.SelectedIndexChanged += new EventHandler(settingsChanged);
 
             this.pictureBoxSSHClientPathWarningIcon.Visible = false;
             this.pictureBoxWinScpPathWarningIcon.Visible = false;
@@ -115,6 +129,7 @@ namespace QuickConnectPlugin {
             this.settings.WinScpPath = this.textBoxWinScpPath.Text;
             this.settings.HostAddressMapFieldName = (String)this.comboBoxHostAddressMapFieldName.SelectedItem;
             this.settings.ConnectionMethodMapFieldName = (String)this.comboBoxConnectionMethodMapFieldName.SelectedItem;
+            this.settings.AdditionalOptionsMapFieldName = (String)this.comboBoxAdditionalOptionsMapFieldName.SelectedItem;
             this.settings.Save();
         }