+++ QuickConnectPlugin/HostPwEntry.cs
@@ -5,13 +5,13 @@ using KeePassLib;
 
 namespace QuickConnectPlugin {
 
-    public class HostPwEntry {
+    public class HostPwEntry : IHostPwEntry {
 
         private String ipAddress;
         public String IPAddress {
             get {
                 if (this.ipAddress == null) {
-                    this.ipAddress = this.pwEntry.Strings.ReadSafe(this.hostAddressFieldName);
+                    this.ipAddress = this.pwEntry.Strings.ReadSafe(this.fieldsMapper.HostAddress);
                 }
                 return ipAddress;
             }
@@ -21,23 +21,31 @@ namespace QuickConnectPlugin {
         public ICollection<ConnectionMethodType> ConnectionMethods {
             get {
                 if (this.connectionMethods == null) {
-                    var value = this.pwEntry.Strings.ReadSafe(this.connectionMethodFieldName);
+                    var value = this.pwEntry.Strings.ReadSafe(this.fieldsMapper.ConnectionMethod);
                     this.connectionMethods = new List<ConnectionMethodType>(ConnectionMethodTypeUtils.GetConnectionMethodsFromString(value));
                 }
                 return new Collection<ConnectionMethodType>(this.connectionMethods);
             }
         }
 
+        private string additionalOptions = null;
+        public string AdditionalOptions {
+            get {
+                if (this.additionalOptions == null) {
+                    this.additionalOptions = this.pwEntry.Strings.ReadSafe(this.fieldsMapper.AdditionalOptions);
+                }
+                return this.additionalOptions;
+            }
+        }
+
         private PwEntry pwEntry;
         private PwDatabase pwDatabase;
-        private String connectionMethodFieldName;
-        private String hostAddressFieldName;
+        private IFieldMapper fieldsMapper;
 
-        public HostPwEntry(PwEntry pwEntry, PwDatabase pwDatabase, String connectionMethodFieldName, String hostAddressFieldName) {
+        public HostPwEntry(PwEntry pwEntry, PwDatabase pwDatabase, IFieldMapper fieldsMapper) {
             this.pwEntry = pwEntry;
             this.pwDatabase = pwDatabase;
-            this.connectionMethodFieldName = connectionMethodFieldName;
-            this.hostAddressFieldName = hostAddressFieldName;
+            this.fieldsMapper = fieldsMapper;
         }
 
         public bool HasIPAddress {