+++ QuickConnectPlugin/IFieldMapper.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace QuickConnectPlugin {
+
+    public interface IFieldMapper {
+
+        String HostAddress { get; }
+        String ConnectionMethod { get; }
+        String AdditionalOptions { get; }
+    }
+}